test_requirements.txt
pytest>=2.8.2
pytest-django>=2.9.1
